added a step-by-step workthrough as deployment guide
The documentation is stored in a new folder `docu/HowToInstall`. The repository README links to the README in the new folder.

The explaination starts with the repository cloning, continues with some configuration workload in Xcode and finishes with the app on the iPhone. Many steps are explained with screenshots taken from a real workthrough.

closes agrinman/sift-ios#7